fix yaml `!!merge` usage with duplicate keys
If we have:
```yaml
defs:
  - &mydef {a: 1, b: 2}
stuff:
  <<: *mydef
  a:
```
we should end up with `a: !!null` and `b: 2` where `a: 1` is
skipped from the mydef merge.  This is consistent with how yaml.Unmarshal
works with duplication keys with !!merge.  The fix is upstream in walky.

This also fixes an issue with backwards compat changes for arrays.
Arrays should be allowed to have duplicate values if those dups are
coming from the original array source, and not from extra merged in
configs.

Also fixed is an edge-case where some options were not being seen
as empty when derived from `!!null` yaml fields.